+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # 0.2.0
 - Added CharacterData#stars and LightConeData#stars.
 - Added LightConeData#superimpositions.
-- Added relics structures and StarRail#getAllRelics.
+- Added relic structures and StarRail#getAllRelics.
 # 0.1.0
 - Released the first version.